[backend] Import map metadata (again)
This commit allows clients to look at metadata for a given map ID. We no
longer throw all the metadata away when doing the conversion, we also
keep it around in a separate key indexed the same way as the map data.
This meant we had to change laps.mapdata into laps.mapdata.image.